Added third (composite) puzzle
This puzzle builds upon the first two.  It also should directly test the
companion on what path was chosen for puzzle 2 (partyhood2) as well as
the first and last portions of this puzzle which test puzzle 1
(partyhood1).

The solution is to get the white for the lower-right guard, then get the
combination for either of the two on the left or right middles, then
depending on which you chose to use that to get through the last gate.

This currently is buggy for the companion for an unknown reason.